proper spice conversion
